# openapi_agent_workflow_platform_mvp_v1.yaml
openapi: 3.0.3
info:
  title: Agent Workflow Platform MVP API
  version: 0.1.0
  description: MVP API draft for workflow authoring, publishing, running, tracing, knowledge bases, tools, models, and secrets.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /workflows:
    post:
      tags: [Workflows]
      summary: Create workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowRequest"
      responses:
        "200":
          description: Created workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Workflows]
      summary: List workflows
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, published, archived]
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Workflow list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /workflows/{workflow_id}:
    get:
      tags: [Workflows]
      summary: Get workflow detail
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      responses:
        "200":
          description: Workflow detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags: [Workflows]
      summary: Update workflow draft
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowRequest"
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Workflows]
      summary: Soft delete workflow
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /workflows/{workflow_id}/validate:
    post:
      tags: [Workflows]
      summary: Validate workflow graph
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateGraphRequest"
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphValidationResult"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /workflows/{workflow_id}/publish:
    post:
      tags: [Workflows]
      summary: Publish workflow version
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishWorkflowRequest"
      responses:
        "200":
          description: Published version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishWorkflowResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /workflows/{workflow_id}/versions:
    get:
      tags: [Workflow Versions]
      summary: List workflow versions
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Version list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersionListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /workflows/{workflow_id}/run:
    post:
      tags: [Runs]
      summary: Run workflow
      parameters:
        - $ref: "#/components/parameters/WorkflowId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunWorkflowRequest"
      responses:
        "200":
          description: Run result or async run id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunWorkflowResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /tools:
    post:
      tags: [Tools]
      summary: Create tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolRequest"
      responses:
        "200":
          description: Tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Tools]
      summary: List tools
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: type
          in: query
          schema:
            type: string
            enum: [api]
      responses:
        "200":
          description: Tool list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /tools/{tool_id}:
    get:
      tags: [Tools]
      summary: Get tool detail
      parameters:
        - $ref: "#/components/parameters/ToolId"
      responses:
        "200":
          description: Tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        default:
          $ref: "#/components/responses/ErrorResponse"
    put:
      tags: [Tools]
      summary: Update tool
      parameters:
        - $ref: "#/components/parameters/ToolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateToolRequest"
      responses:
        "200":
          description: Tool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tool"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /tools/{tool_id}/test:
    post:
      tags: [Tools]
      summary: Test tool
      parameters:
        - $ref: "#/components/parameters/ToolId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TestToolRequest"
      responses:
        "200":
          description: Tool test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TestToolResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    WorkflowId:
      name: workflow_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    ToolId:
      name: tool_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
    Timestamp:
      type: string
      format: date-time
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              $ref: "#/components/schemas/JsonObject"
        request_id:
          type: string
    WorkflowStatus:
      type: string
      enum: [draft, published, archived]
    WorkflowRunStatus:
      type: string
      enum: [pending, running, completed, failed, cancelled]
...
    NodeType:
      type: string
      enum: [start, input, llm, knowledge_base, intent, branch, api, message, output, end]
    WorkflowGraph:
      type: object
      required: [schema_version, nodes, edges]
      properties:
        schema_version:
          type: string
          example: "1.0"
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowEdge"
    WorkflowNode:
      type: object
      required: [id, type, name, position, config]
      properties:
        id:
          type: string
          example: llm_1
        type:
          $ref: "#/components/schemas/NodeType"
        name:
          type: string
        description:
          type: string
        position:
          type: object
          required: [x, y]
          properties:
            x:
              type: number
            y:
              type: number
        input_mapping:
          $ref: "#/components/schemas/JsonObject"
        output_mapping:
          $ref: "#/components/schemas/JsonObject"
        config:
          $ref: "#/components/schemas/JsonObject"
        retry:
          $ref: "#/components/schemas/RetryPolicy"
        timeout:
          type: integer
          example: 60
        on_error:
          $ref: "#/components/schemas/OnErrorPolicy"
        enabled:
          type: boolean
          default: true
    WorkflowEdge:
      type: object
      required: [id, source, target]
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        label:
          type: string
        condition:
          nullable: true
          oneOf:
            - type: string
            - $ref: "#/components/schemas/JsonObject"
    RetryPolicy:
      type: object
      properties:
        max_attempts:
          type: integer
          minimum: 1
          default: 1
        backoff:
          type: string
          enum: [none, fixed, exponential]
          default: none
        retry_on:
          type: array
          items:
            type: string
    OnErrorPolicy:
      type: object
      properties:
        strategy:
          type: string
          enum: [fail_workflow, skip_node, go_to_node]
          default: fail_workflow
        target:
          type: string
    CreateWorkflowRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        draft_graph_json:
          $ref: "#/components/schemas/WorkflowGraph"
    UpdateWorkflowRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        draft_graph_json:
          $ref: "#/components/schemas/WorkflowGraph"
    WorkflowDetail:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          $ref: "#/components/schemas/WorkflowStatus"
        current_version_id:
          type: integer
          format: int64
          nullable: true
        draft_graph_json:
          $ref: "#/components/schemas/WorkflowGraph"
        created_by:
          type: integer
          format: int64
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    WorkflowListItem:
      allOf:
        - $ref: "#/components/schemas/WorkflowDetail"
        - type: object
          properties:
            current_version:
              type: integer
              nullable: true
            latest_run:
              type: object
              nullable: true
              properties:
                run_id:
                  type: integer
                  format: int64
                status:
                  $ref: "#/components/schemas/WorkflowRunStatus"
                created_at:
                  $ref: "#/components/schemas/Timestamp"
    WorkflowListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowListItem"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    ValidateGraphRequest:
      type: object
      required: [mode, graph_json]
      properties:
        mode:
          type: string
          enum: [draft, publish, run]
        graph_json:
          $ref: "#/components/schemas/WorkflowGraph"
    GraphValidationResult:
      type: object
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/GraphValidationIssue"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/GraphValidationIssue"
    GraphValidationIssue:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        path:
          type: string
        node_id:
          type: string
          nullable: true
    PublishWorkflowRequest:
      type: object
      properties:
        release_note:
          type: string
    PublishWorkflowResponse:
      type: object
      properties:
        workflow_id:
          type: integer
          format: int64
        version_id:
          type: integer
          format: int64
        version:
          type: integer
        schema_version:
          type: string
        created_at:
          $ref: "#/components/schemas/Timestamp"
    WorkflowVersionSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        workflow_id:
          type: integer
          format: int64
        version:
          type: integer
        schema_version:
          type: string
        release_note:
          type: string
        published_by:
          type: integer
          format: int64
        created_at:
          $ref: "#/components/schemas/Timestamp"
...
    WorkflowVersionListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/WorkflowVersionSummary"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    RunWorkflowRequest:
      type: object
      required: [input]
      properties:
        input:
          $ref: "#/components/schemas/JsonObject"
        version_id:
          type: integer
          format: int64
        trigger_type:
          type: string
          enum: [manual, api, test]
          default: manual
        execution_mode:
          type: string
          enum: [sync, async]
          default: sync
    RunWorkflowResponse:
      type: object
      properties:
        run_id:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/WorkflowRunStatus"
        output:
          $ref: "#/components/schemas/JsonObject"
        started_at:
          $ref: "#/components/schemas/Timestamp"
        ended_at:
          $ref: "#/components/schemas/Timestamp"
...
    Tool:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
          enum: [api]
        description:
          type: string
        config_json:
          $ref: "#/components/schemas/JsonObject"
        status:
          type: string
        created_at:
          $ref: "#/components/schemas/Timestamp"
    CreateToolRequest:
      type: object
      required: [name, type, config]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [api]
        description:
          type: string
        config:
          $ref: "#/components/schemas/JsonObject"
    ToolListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Tool"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    TestToolRequest:
      type: object
      properties:
        input:
          $ref: "#/components/schemas/JsonObject"
    TestToolResponse:
      type: object
      properties:
        success:
          type: boolean
        status_code:
          type: integer
        duration_ms:
          type: integer
        response:
          nullable: true
          oneOf:
            - type: object
              additionalProperties: true
            - type: array
              items: {}
            - type: string
        error_message:
          type: string
          nullable: true
...